Fix: Make license parsing namespace-aware (#49)
The XML parser was previously using a hardcoded `xlink` prefix to extract the `href` attribute from the `<license>` tag. This would fail if the XML used a different prefix for the `http://www.w3.org/1999/xlink` namespace.

This commit fixes the issue by using the element's namespace map (`nsmap`) to correctly resolve the `href` attribute regardless of the prefix used.

A new unit test has been added to verify this fix. The test uses a sample XML file with a custom namespace prefix to ensure the parser can handle this variation correctly.